# Faster campaigns tests, part 2: determine plan without a database.